Transition Commit
In this commit I was setting up the transition for the navigations hypertext links, and I added in the ba_target.png to be the background of the navigation list. This process took a while, but I got some help and found out that I needed to add in the background-size in order to transition the image.